+++ src/Presentation/SmartStore.Web.Framework/Plugins/PluginHelperFeed.cs
@@ -12,9 +12,9 @@ using SmartStore.Services.Catalog;
 using SmartStore.Services.Media;
 using SmartStore.Services.Tasks;
 using SmartStore.Services.Stores;
-using SmartStore.Services.Seo;
 using SmartStore.Core.Domain.Stores;
 using SmartStore.Web.Framework.Mvc;
+using SmartStore.Services.Seo;
 
 namespace SmartStore.Web.Framework.Plugins
 {